Renamed role details files
