+++ frontend/src/routes/UserManagement/Roles/Role/RoleDetails.test.tsx
@@ -4,7 +4,7 @@ import { render, screen } from '@testing-library/react'
 import { MemoryRouter } from 'react-router-dom-v5-compat'
 import { RecoilRoot } from 'recoil'
 import { nockIgnoreRBAC } from '../../../../lib/nock-util'
-import { RoleDetail } from './RoleDetail'
+import { RoleDetail } from './RoleDetails'
 
 function Component({ roleId = 'test-role' }: { roleId?: string }) {
   return (
