+++ frontend/src/routes/UserManagement/Roles/RolesManagement.tsx
@@ -3,7 +3,7 @@ import { Navigate, Route, Routes } from 'react-router-dom-v5-compat'
 import { NavigationPath, createRoutePathFunction } from '../../../NavigationPath'
 import { RolesPage, RolesContextProvider } from './RolesPage'
 import { RolePage } from './Role/RolePage'
-import { RoleDetail } from './Role/RoleDetail'
+import { RoleDetail } from './Role/RoleDetails'
 import { RoleYaml } from './Role/RoleYaml'
 import { RolePermissions } from './Role/RolePermissions'
 import { RoleAssignments } from './Role/RoleAssignments'
